requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
ecdsa==0.19.1
fastapi==0.110.0
h11==0.16.0
idna==3.11
passlib==1.7.4
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
python-dotenv==1.2.2
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9.1
six==1.17.0
sqlalchemy==2.0.29
starlette==0.36.3
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.29.0
